+++ src/flows/shared/steps/OrganizeCollection/Editor/StagedNftImage.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { EditModeNft } from 'types/Nft';
+import { EditModeNft } from 'hooks/api/nfts/types';
 
 import NftPreviewLabel from 'components/NftPreview/NftPreviewLabel';
 import transitions from 'components/core/transitions';
