+++ src/hooks/api/galleries/types.ts
@@ -1,0 +1,3 @@
+import { Gallery } from 'types/Gallery';
+
+export type GetGalleriesResponse = { galleries: Gallery[] };
