+++ src/hooks/api/nfts/types.ts
@@ -1,0 +1,8 @@
+import { Nft } from 'types/Nft';
+
+export type EditModeNft = {
+  id: string;
+  nft: Nft;
+  isSelected?: boolean;
+  index: number;
+};
