+++ src/hooks/api/users/types.ts
@@ -1,0 +1,7 @@
+export type UpdateUserResponse = null;
+
+export type UpdateUserRequest = {
+  user_id: string;
+  username: string;
+  description: string;
+};
