+++ src/hooks/api/users/useUser.ts
@@ -2,7 +2,7 @@ import { useAuthState } from 'contexts/auth/AuthContext';
 import { isLoggedInState } from 'contexts/auth/types';
 import { useMemo } from 'react';
 import { User } from 'types/User';
-import useGet from './rest/useGet';
+import useGet from '../_rest/useGet';
 
 type Props = {
   id?: string | null;
