+++ src/scenes/NftDetailPage/NftDetailAsset.tsx
@@ -3,12 +3,11 @@ import { NftMediaType } from 'components/core/enums';
 import { useMemo } from 'react';
 import styled from 'styled-components';
 
-import { Nft } from 'types/Nft';
-
 import NftDetailAnimation from './NftDetailAnimation';
 import NftDetailVideo from './NftDetailVideo';
 import NftDetailAudio from './NftDetailAudio';
 import ImageWithLoading from 'components/ImageWithLoading/ImageWithLoading';
+import { Nft } from 'types/Nft';
 
 type Props = {
   nft: Nft;
