+++ src/scenes/UserGalleryPage/UserGallery.tsx
@@ -1,14 +1,11 @@
-import breakpoints, {
-  contentSize,
-  pageGutter,
-} from 'components/core/breakpoints';
+import { contentSize } from 'components/core/breakpoints';
 import styled from 'styled-components';
+import Spacer from 'components/core/Spacer/Spacer';
 import { Gallery } from 'types/Gallery';
-import { User } from 'types/User';
 import UserGalleryCollections from './UserGalleryCollections';
 import UserGalleryHeader from './UserGalleryHeader';
 import EmptyGallery from './EmptyGallery';
-import Spacer from 'components/core/Spacer/Spacer';
+import { User } from 'types/User';
 
 type Props = {
   user: User;
