+++ src/types/Collection.ts
@@ -7,22 +7,3 @@ export type Collection = {
   description?: string;
   isHidden?: boolean;
 };
-
-export type CreateCollectionRequest = {
-  // list of NFT IDs
-  nfts: string[];
-};
-
-export type CreateCollectionResponse = {
-  collection_id: string;
-};
-
-export type GetCollectionsResponse = { collections: Collection[] };
-
-export type UpdateCollectionRequest = {
-  id: string;
-  name: string;
-  collectors_note: string;
-};
-
-export type UpdateCollectionResponse = null;
