+++ src/types/User.ts
@@ -6,11 +6,3 @@ export type User = {
   description: string;
   addresses: string[];
 };
-
-export type UpdateUserResponse = null;
-
-export type UpdateUserRequest = {
-  user_id: string;
-  username: string;
-  description: string;
-};
